tests written and interfaces slightly modified

+++ src/interfaces/IRouter.sol
@@ -3,10 +3,10 @@ pragma solidity ^0.8.20;
 
 interface IRouter {
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory);
+        uint256 deadline
+    ) external returns (uint256[] memory);
 }
